requirements.txt
certifi==2019.6.16
Chameleon==3.6.2
chardet==3.0.4
hupper==1.8.1
idna==2.8
Jinja2==2.10.1
Mako==1.0.13
MarkupSafe==1.1.1
PasteDeploy==2.0.1
plaster==1.0
plaster-pastedeploy==0.7
Pygments==2.4.2
pyramid==1.10.4
pyramid-chameleon==0.3
pyramid-debugtoolbar==4.5
pyramid-jinja2==2.8
pyramid-mako==1.0.2
-e git+https://vitorsgobbi@bitbucket.org/vitorsgobbi/quotes.git@3e19ba4b91b12841d55ba71b2e0efab2e9480da4#egg=quotes
repoze.lru==0.7
requests==2.22.0
SQLAlchemy==1.3.5
transaction==2.4.0
translationstring==1.3
urllib3==1.25.3
venusian==1.2.0
waitress==1.3.0
WebOb==1.8.5
zope.deprecation==4.4.0
zope.interface==4.6.0
zope.sqlalchemy==1.1
